refactor: move parsing and color helpers
