+++ app_components/plotting.py
@@ -1,8 +1,4 @@
-import json
-from collections import defaultdict
 from datetime import timedelta
-from math import floor
-from pathlib import Path
 
 import pandas as pd
 import plotly.graph_objs as go
@@ -16,131 +12,6 @@ def get_layout_config(screen_width):
     hour_height = 20 if screen_width < 480 else 36 if screen_width < 768 else 44
     label_column_pct = 10
     return hour_height, label_column_pct
-
-
-# Color map by Casino
-DATA_DIR = Path(__file__).resolve().parent.parent / "data"
-with open(DATA_DIR / "casino_colors.json", encoding="utf-8") as f:
-    COLOR_MAP = json.load(f)
-with open(DATA_DIR / "default_colors.json", encoding="utf-8") as f:
-    DEFAULT_COLORS = json.load(f)
-
-
-def get_color():
-    color_map = COLOR_MAP
-    default_colors = DEFAULT_COLORS
-
-    result = {}
-    for casino, colors in color_map.items():
-        result[casino] = colors
-
-    if not result:
-        dummy_casinos = [f"Casino {i}" for i in range(len(default_colors))]
-        for casino_name, color in zip(dummy_casinos, default_colors):
-            result[casino_name] = {"bg": color, "text": "#000000"}
-
-    return result
-
-
-# Add arrow indicators for events that span outside the week
-def annotate_events_with_flags(events_df, week_start, week_end):
-    events_df = events_df.copy()
-    # Preserve the original index so CSS grid clicks can reference the global row
-    events_df["orig_index"] = events_df.index
-
-    # Add a duration column for sorting, and sort by: both left and right arrows, only left arrow, fully within week, and only right arrow
-    events_df["Duration"] = (
-        events_df["EndDate"] - events_df["StartDate"]
-    ).dt.total_seconds()
-    events_df["has_left_arrow"] = events_df["StartDate"] < week_start
-    events_df["has_right_arrow"] = events_df["EndDate"] > week_end
-
-    def get_overflow_priority(row):
-        # Overflow priority: both arrows -> 0, right only -> 1, none -> 2, left only -> 3
-        if row["has_left_arrow"] and row["has_right_arrow"]:
-            return 0
-        if row["has_right_arrow"]:
-            return 3
-        if not row["has_left_arrow"] and not row["has_right_arrow"]:
-            return 2
-        return 1  # left only
-
-    events_df["overflow_sort"] = events_df.apply(get_overflow_priority, axis=1)
-
-    # Sort events and drop the previous index, but keep the preserved orig_index column
-    return events_df.sort_values(
-        by=["overflow_sort", "StartDate", "EndDate", "Duration", "Casino"],
-        ascending=[True, True, True, False, True],
-    ).reset_index(drop=True)
-
-
-# Filter events that overlap with the current week, excluding long_spanning events
-def filter_week_events(events_df, week_start, week_end):
-    return events_df[
-        (events_df["EndDate"] > week_start)
-        & (events_df["StartDate"] < week_end)
-        & ~(events_df["StartDate"] == week_end)
-        & ~((events_df["StartDate"] < week_start) & (events_df["EndDate"] > week_end))
-    ].copy()
-
-
-def assign_event_rows(events_df, week_start):
-    # Layout params
-    used_rows_by_day = {i: set() for i in range(7)}
-    recurring_rows = defaultdict(int)
-    current_row = 0
-    row_nums = []
-
-    for priority in sorted(events_df["overflow_sort"].unique()):
-        group_df = events_df[events_df["overflow_sort"] == priority].sort_values(
-            by=["StartDate", "EndDate", "Duration", "Casino"],
-            ascending=[True, True, False, True],
-        )
-
-        for idx, row in group_df.iterrows():
-            row = events_df.loc[idx]
-            start_delta = (row["StartDate"] - week_start).total_seconds() / (24 * 3600)
-            end_delta = (row["EndDate"] - week_start).total_seconds() / (24 * 3600)
-
-            # Calculate the visible range of the event within the 7-day week
-            visible_start = max(start_delta, 0)
-            visible_end = min(end_delta, 7)
-
-            start_day = max(0, floor(visible_start))
-            end_day = min(6, floor(visible_end - 1e-6))
-
-            recurring_key = f"{row['EventName']}|{row['Casino']}|{row['StartDate'].time()}|{row['EndDate'].time()}"
-            preferred_row = recurring_rows.get(recurring_key)
-            row_assigned = False
-
-            # First try preferred row
-            if preferred_row is not None and all(
-                preferred_row not in used_rows_by_day[d]
-                for d in range(start_day, end_day + 1)
-            ):
-                assigned_row = preferred_row
-                row_assigned = True
-            else:
-                for r in range(current_row, 100):
-                    if all(
-                        r not in used_rows_by_day[d]
-                        for d in range(start_day, end_day + 1)
-                    ):
-                        assigned_row = r
-                        recurring_rows[recurring_key] = r
-                        row_assigned = True
-                        break
-
-            # If not usable, find a new row
-            if row_assigned:
-                for d in range(start_day, end_day + 1):
-                    used_rows_by_day[d].add(assigned_row)
-                events_df.at[idx, "row_num"] = assigned_row
-                row_nums.append(assigned_row)
-
-        current_row = max(row_nums, default=current_row) + 1
-
-    return events_df
 
 
 # Generate a responsive 24-hour vertical day view with absolutely positioned event blocks.
