+++ app_components/week_grid_layout.py
@@ -1,12 +1,13 @@
 import pandas as pd
 from dash import html
 
-from .plotting import (
+from utils.colors import get_color
+from utils.data_parsing import (
     annotate_events_with_flags,
     assign_event_rows,
     filter_week_events,
-    get_color,
 )
+
 from .utils import get_week_range, trim_label
