+++ utils/__init__.py
@@ -1,0 +1,13 @@
+from .colors import get_color
+from .data_parsing import (
+    annotate_events_with_flags,
+    assign_event_rows,
+    filter_week_events,
+)
+
+__all__ = [
+    "get_color",
+    "annotate_events_with_flags",
+    "filter_week_events",
+    "assign_event_rows",
+]
